fix: move guide runtime to egypt-reachable host
