+++ .github/workflows/automated-release-blog-post.yml
@@ -347,7 +347,7 @@ jobs:
           - ⭐ [Star SHAFT on GitHub](https://github.com/ShaftHQ/SHAFT_ENGINE) — it helps more than you think
           - 💡 [Start a Discussion](https://github.com/ShaftHQ/SHAFT_ENGINE/discussions) — share feedback or ideas
           - 🐛 [Report an Issue](https://github.com/ShaftHQ/SHAFT_ENGINE/issues/new) — help us squash bugs faster
-          - 📖 [Read the Docs](https://shaftengine.netlify.app/docs/start/overview) — dive deeper into SHAFT
+          - 📖 [Read the Docs](https://shaft-engine.automatest.org/docs/start/overview) — dive deeper into SHAFT
 
           Thanks for being part of the SHAFT journey. Until the next release — happy testing! 🎯
           `
