+++ .github/workflows/deploy.yml
@@ -1,4 +1,4 @@
-name: Deploy Redirect to GitHub Pages
+name: Deploy Static Guide to GitHub Pages
 
 on:
   workflow_dispatch:
@@ -11,7 +11,7 @@ on:
 
 jobs:
   deploy:
-    name: Deploy Redirect to GitHub Pages
+    name: Deploy Static Guide to GitHub Pages
     runs-on: ubuntu-latest
     permissions:
       contents: write  # Required to push to gh-pages branch
@@ -25,16 +25,17 @@ jobs:
       - name: Install
         run: yarn install
 
-      # Deploy redirect page to GitHub Pages
-      # The redirect-page directory contains a simple HTML redirect to https://shaftengine.netlify.app/
-      # Netlify is now the primary deployment platform, GitHub Pages only serves as a redirect
+      - name: Build
+        run: yarn build
+
+      # Publish the full static guide to GitHub Pages as a static-only fallback.
+      # AutoBot still requires the Cloudflare Pages Function on the canonical domain.
       # Docs: https://github.com/peaceiris/actions-gh-pages#%EF%B8%8F-docusaurus
       - name: Release
         uses: peaceiris/actions-gh-pages@84c30a85c19949d7eee79c4ff27748b70285e453 # v4.1.0
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
-          # Publish only the redirect page to the `gh-pages` branch:
-          publish_dir: ./redirect-page
+          publish_dir: ./build
           # The following lines assign commit authorship to the official
           # GH-Actions bot for deploys to `gh-pages` branch:
           # https://github.com/actions/checkout/issues/13#issuecomment-724415212
