+++ blog/2026-03-24-release-10.1.20260324.md
@@ -105,6 +105,6 @@ We'd love to hear what you think! Found a bug? Have an idea? Just want to say hi
 - ⭐ [Star SHAFT on GitHub](https://github.com/ShaftHQ/SHAFT_ENGINE) — it helps more than you think
 - 💡 [Start a Discussion](https://github.com/ShaftHQ/SHAFT_ENGINE/discussions) — share feedback or ideas
 - 🐛 [Report an Issue](https://github.com/ShaftHQ/SHAFT_ENGINE/issues/new) — help us squash bugs faster
-- 📖 [Read the Docs](https://shafthq.github.io/) — dive deeper into SHAFT
+- 📖 [Read the Docs](https://shaft-engine.automatest.org/) — dive deeper into SHAFT
 
 Thanks for being part of the SHAFT journey. Until the next release — happy testing! 🎯
