+++ blog/2026-05-02-release-10.2.20260501.md
@@ -75,7 +75,7 @@ Or update the dependency directly:
 | Resource | Link |
 |---|---|
 | 📦 Release Notes | [10.2.20260501](https://github.com/ShaftHQ/SHAFT_ENGINE/releases/tag/10.2.20260501) |
-| 📚 Full Documentation | [shafthq.github.io](https://shafthq.github.io/) |
+| 📚 Full Documentation | [shaft-engine.automatest.org](https://shaft-engine.automatest.org/) |
 | 💬 Community | [GitHub Discussions](https://github.com/ShaftHQ/SHAFT_ENGINE/discussions) |
 | 🐛 Report an Issue | [Issue Tracker](https://github.com/ShaftHQ/SHAFT_ENGINE/issues/new) |
 
@@ -189,6 +189,6 @@ We'd love to hear what you think! Found a bug? Have an idea? Just want to say hi
 - ⭐ [Star SHAFT on GitHub](https://github.com/ShaftHQ/SHAFT_ENGINE) — it helps more than you think
 - 💡 [Start a Discussion](https://github.com/ShaftHQ/SHAFT_ENGINE/discussions) — share feedback or ideas
 - 🐛 [Report an Issue](https://github.com/ShaftHQ/SHAFT_ENGINE/issues/new) — help us squash bugs faster
-- 📖 [Read the Docs](https://shafthq.github.io/) — dive deeper into SHAFT
+- 📖 [Read the Docs](https://shaft-engine.automatest.org/) — dive deeper into SHAFT
 
 Thanks for being part of the SHAFT journey. Until the next release — happy testing! 🎯
