+++ blog/2026-05-06-release-10.2.20260506.md
@@ -76,8 +76,8 @@ Or update the dependency directly:
 
 | Resource | Link |
 |---|---|
-| 📚 Full Documentation | [shafthq.github.io](https://shafthq.github.io/) |
-| 🚀 Getting Started | [Quick-Start Guide](https://shafthq.github.io/) |
+| 📚 Full Documentation | [shaft-engine.automatest.org](https://shaft-engine.automatest.org/) |
+| 🚀 Getting Started | [Quick-Start Guide](https://shaft-engine.automatest.org/) |
 | 📋 JavaDocs | [ShaftHQ JavaDoc](https://shafthq.github.io/SHAFT_ENGINE/) |
 | 🤖 MCP Server | [shaft-mcp](https://github.com/ShaftHQ/shaft-mcp) |
 | 🗺️ Roadmap | [GitHub Projects](https://github.com/orgs/ShaftHQ/projects) |
@@ -187,6 +187,6 @@ We'd love to hear what you think! Found a bug? Have an idea? Just want to say hi
 - ⭐ [Star SHAFT on GitHub](https://github.com/ShaftHQ/SHAFT_ENGINE) — it helps more than you think
 - 💡 [Start a Discussion](https://github.com/ShaftHQ/SHAFT_ENGINE/discussions) — share feedback or ideas
 - 🐛 [Report an Issue](https://github.com/ShaftHQ/SHAFT_ENGINE/issues/new) — help us squash bugs faster
-- 📖 [Read the Docs](https://shafthq.github.io/) — dive deeper into SHAFT
+- 📖 [Read the Docs](https://shaft-engine.automatest.org/) — dive deeper into SHAFT
 
 Thanks for being part of the SHAFT journey. Until the next release — happy testing! 🎯
