+++ blog/2026-06-22-release-10.2.20260622.md
@@ -27,7 +27,7 @@ A fresh release just dropped and it's packed with improvements you don't want to
 
 Auto-generated release notes are included below. Users upgrading from
 `SHAFT_ENGINE` should read the
-[modular SHAFT upgrade guide](https://shaftengine.netlify.app/docs/start/upgrade).
+[modular SHAFT upgrade guide](https://shaft-engine.automatest.org/docs/start/upgrade).
 
 SHAFT Pilot adds deterministic Capture, TestNG generation, Doctor diagnosis,
 reviewed repair proposals, and MCP interoperability. AI is optional, disabled
@@ -35,7 +35,7 @@ by default, and direct OpenAI, Anthropic, Gemini, or Ollama access requires
 explicit enablement and consent. Microsoft/GitHub Copilot integrates through
 MCP rather than a generic provider API-key adapter.
 
-See the [SHAFT Pilot guide](https://shaftengine.netlify.app/docs/agentic/pilot)
+See the [SHAFT Pilot guide](https://shaft-engine.automatest.org/docs/agentic/pilot)
 for installation, configuration, privacy, troubleshooting, and usage examples.
 
 <!-- Release notes generated using configuration in .github/release.yml at main -->
@@ -134,6 +134,6 @@ We'd love to hear what you think! Found a bug? Have an idea? Just want to say hi
 - ⭐ [Star SHAFT on GitHub](https://github.com/ShaftHQ/SHAFT_ENGINE) — it helps more than you think
 - 💡 [Start a Discussion](https://github.com/ShaftHQ/SHAFT_ENGINE/discussions) — share feedback or ideas
 - 🐛 [Report an Issue](https://github.com/ShaftHQ/SHAFT_ENGINE/issues/new) — help us squash bugs faster
-- 📖 [Read the Docs](https://shaftengine.netlify.app/docs/start/overview) — dive deeper into SHAFT
+- 📖 [Read the Docs](https://shaft-engine.automatest.org/docs/start/overview) — dive deeper into SHAFT
 
 Thanks for being part of the SHAFT journey. Until the next release — happy testing! 🎯
