+++ docs/archive/site/netlify-migration-guide.md
@@ -6,7 +6,9 @@ tags: [archive]
 
 # Migration Guide: GitHub Pages to Netlify
 
-> **Note:** This migration has been completed! The site is now fully deployed to Netlify at https://shaftengine.netlify.app/, and GitHub Pages (https://shafthq.github.io/) automatically redirects all visitors to the Netlify deployment.
+> **Note:** This guide records a previous migration to Netlify. The current
+> canonical deployment is described in
+> [Documentation site operations](/docs/maintainers/site-operations).
 
 ## Overview
 
@@ -183,7 +185,7 @@ To ensure users visiting the old GitHub Pages URL are automatically redirected t
 
 1. A simple redirect page is maintained in the `redirect-page/` directory
 2. The GitHub Actions workflow (`.github/workflows/deploy.yml`) deploys only this redirect page to GitHub Pages
-3. When users visit https://shafthq.github.io/, they are automatically redirected to https://shaftengine.netlify.app/
+3. When users visit the old GitHub Pages URL, they are automatically redirected to the Netlify deployment
 4. The redirect uses both meta refresh and JavaScript for maximum compatibility
 
 This ensures a seamless transition for users who have bookmarked or linked to the old URL.
