+++ netlify/functions/docs-loader.mjs
@@ -2,25 +2,17 @@ import {readFileSync, readdirSync} from 'fs';
 import {dirname, extname, join, relative, sep} from 'path';
 import {fileURLToPath} from 'url';
 import {
-  SHAFT_GITHUB_DISCUSSIONS,
-  SHAFT_GITHUB_ISSUES,
-  SHAFT_GITHUB_ORG,
-  SHAFT_GITHUB_REPO,
-} from './constants.mjs';
+  loadDocumentationFromIndex,
+  retrieveFromIndex,
+} from './docs-retrieval.mjs';
+export { getGitHubRepositoryContext } from './github-context.mjs';
 
 const currentDir = dirname(fileURLToPath(import.meta.url));
 const snippets = JSON.parse(
   readFileSync(join(currentDir, '..', '..', 'src', 'data', 'snippets.json'), 'utf-8'),
 );
 const MCP_COMMAND_SYSTEMS = ['windows', 'macos', 'linux'];
-const MAX_RETRIEVAL_CHARACTERS = 80_000;
-const MAX_RETRIEVAL_CHUNKS = 8;
 const EXCLUDED_DIRECTORIES = new Set(['archive', 'maintainers']);
-const STOP_WORDS = new Set([
-  'a', 'an', 'and', 'are', 'as', 'at', 'be', 'by', 'for', 'from', 'how',
-  'in', 'is', 'it', 'of', 'on', 'or', 'that', 'the', 'this', 'to', 'use',
-  'what', 'when', 'where', 'which', 'with', 'your',
-]);
 
 let cachedIndex = null;
 
@@ -72,18 +64,6 @@ function normalizeMarkdown(content) {
     .trim();
 }
 
-function tokenize(value) {
-  return [
-    ...new Set(
-      value
-        .toLowerCase()
-        .replace(/[^a-z0-9._-]+/gu, ' ')
-        .split(/\s+/u)
-        .filter((token) => token.length > 1 && !STOP_WORDS.has(token)),
-    ),
-  ];
-}
-
 function splitIntoChunks(file) {
   const normalized = normalizeMarkdown(file.content);
   const lines = normalized.split('\n');
@@ -128,73 +108,15 @@ export function buildDocumentationIndex() {
   return cachedIndex;
 }
 
-function scoreChunk(chunk, query, queryTokens) {
-  const pathAndHeading = `${chunk.path} ${chunk.heading}`.toLowerCase();
-  let score = 0;
-
-  if (query && chunk.searchText.includes(query)) score += 30;
-  for (const token of queryTokens) {
-    if (pathAndHeading.includes(token)) score += 8;
-    if (chunk.searchText.includes(token)) score += 2;
-  }
-  if (
-    queryTokens.length > 0
-    && queryTokens.every((token) => chunk.searchText.includes(token))
-  ) {
-    score += 20;
-  }
-
-  if (chunk.path === 'start/overview.mdx') score += 0.5;
-  return score;
-}
-
 export function retrieveDocumentation(query, options = {}) {
-  const maxCharacters = options.maxCharacters ?? MAX_RETRIEVAL_CHARACTERS;
-  const maxChunks = options.maxChunks ?? MAX_RETRIEVAL_CHUNKS;
-  const normalizedQuery = String(query || 'SHAFT overview').trim().toLowerCase();
-  const queryTokens = tokenize(normalizedQuery);
-
-  const ranked = buildDocumentationIndex()
-    .map((chunk) => ({...chunk, score: scoreChunk(chunk, normalizedQuery, queryTokens)}))
-    .filter((chunk) => chunk.score > 0)
-    .sort((left, right) => right.score - left.score || left.path.localeCompare(right.path));
-
-  const selected = [];
-  let characters = 0;
-  for (const chunk of ranked) {
-    const rendered = `## ${chunk.heading}\nSource: /docs/${chunk.path.replace(/\.(?:md|mdx)$/u, '')}\n\n${chunk.content}`;
-    if (selected.length >= maxChunks) break;
-    if (characters + rendered.length > maxCharacters) continue;
-    selected.push({...chunk, rendered});
-    characters += rendered.length;
-  }
-
-  return selected;
+  return retrieveFromIndex(buildDocumentationIndex(), query, options);
 }
 
 export function loadDocumentation(query = 'SHAFT overview') {
   try {
-    const selected = retrieveDocumentation(query);
-    const sections = selected.map((chunk) => chunk.rendered).join('\n\n---\n\n');
-    return `# SHAFT Engine official documentation\n\nRetrieved for: ${query}\n\n${sections}`;
+    return loadDocumentationFromIndex(buildDocumentationIndex(), query);
   } catch (error) {
     console.error('[Documentation Loader] Error loading documentation:', error);
     return null;
   }
-}
-
-export function getGitHubRepositoryContext() {
-  return `
-# SHAFT Engine official GitHub repository
-
-Repository: ${SHAFT_GITHUB_REPO}
-Organization: ${SHAFT_GITHUB_ORG}
-
-Use these official resources when a question is outside the retrieved guide:
-- Issues: ${SHAFT_GITHUB_ISSUES}
-- Discussions: ${SHAFT_GITHUB_DISCUSSIONS}
-- JavaDocs: https://shafthq.github.io/SHAFT_ENGINE/
-
-Never invent implementation details that are absent from the retrieved
-documentation.`;
 }
