--- redirect-page/index.html
@@ -1,66 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="refresh" content="0;url=https://shaftengine.netlify.app/">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Redirecting to SHAFT User Guide</title>
-    <style>
-        body {
-            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, sans-serif;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
-            margin: 0;
-            background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-            color: white;
-            text-align: center;
-        }
-        .container {
-            padding: 2rem;
-        }
-        h1 {
-            font-size: 2rem;
-            margin-bottom: 1rem;
-        }
-        p {
-            font-size: 1.1rem;
-            margin-bottom: 1.5rem;
-        }
-        a {
-            color: #fff;
-            text-decoration: underline;
-            font-weight: bold;
-        }
-        .spinner {
-            border: 4px solid rgba(255, 255, 255, 0.3);
-            border-top: 4px solid white;
-            border-radius: 50%;
-            width: 40px;
-            height: 40px;
-            animation: spin 1s linear infinite;
-            margin: 0 auto;
-        }
-        @keyframes spin {
-            0% { transform: rotate(0deg); }
-            100% { transform: rotate(360deg); }
-        }
-    </style>
-    <script>
-        // Redirect to Netlify site immediately
-        window.location.replace('https://shaftengine.netlify.app/');
-    </script>
-</head>
-<body>
-    <div class="container">
-        <h1>We've Moved!</h1>
-        <p>The SHAFT User Guide has moved to a new location.</p>
-        <p>You will be redirected automatically in a moment...</p>
-        <div class="spinner"></div>
-        <p style="margin-top: 2rem;">
-            If you are not redirected, please click <a href="https://shaftengine.netlify.app/">here</a>.
-        </p>
-    </div>
-</body>
-</html>
